+++ src/waving.hh
@@ -5,7 +5,7 @@
 #include "raylib.h"
 
 // Per tal que, per cada segon, a 44100Hz, n'hi hagui approx 20
-#define FFT_CHUNK_SIZE 2048 
+#define FFT_CHUNK_SIZE 2048
 
 float *extreu_fft_from_samples(float *samples, size_t sample_length, float sampleRate);
 
@@ -13,6 +13,6 @@ float *extreu_fft_from_samples(float *samples, size_t sample_length, float sampl
 // double freq = (double)i * la.sampleRate / (double)N;
 float* extreu_fft_from_wav(Wave* la);
 
-float *which_pitch_is_playing_at_each_time_instance(float *samples,
+double *which_pitch_is_playing_at_each_time_instance(float *samples,
                                                     size_t sample_length,
                                                     float sampleRate);
